revert some unnecessary changes
